requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
python-dotenv==0.19.1
requests==2.26.0
markdown2==2.4.0
werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
scholarly==1.7.11
selenium==4.9.0 